# backend/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"MessageResponse": {
				"description": "Generic success envelope",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"StudentProfileBody": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"studentId": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"faculty": {
						"type": "string"
					},
					"yearOfStudy": {
						"type": "number",
						"format": "double"
					},
					"institution": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"skills": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SupervisorProfileBody": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"institution": {
						"type": "string"
					},
					"specialization": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CompanyProfileBody": {
				"properties": {
					"companyName": {
						"type": "string"
					},
					"industry": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"addressLine1": {
						"type": "string"
					},
					"addressLine2": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"StartConversationBody": {
				"properties": {
					"recipientId": {
						"type": "string",
						"description": "User ID of the person to start a conversation with"
					}
				},
				"required": [
					"recipientId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendMessageBody": {
				"properties": {
					"content": {
						"type": "string"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Internship Monitoring & Connecting System API",
		"version": "1.0.0",
		"description": "Full-stack platform bridging students, companies, and supervisors throughout the internship lifecycle.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
...
		"/api/users/student/me": {
			"get": {
				"operationId": "GetStudentProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get the authenticated student's profile",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateStudentProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Update the authenticated student's profile",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StudentProfileBody"
							}
						}
					}
				}
			}
		},
		"/api/users/student/me/photo": {
			"post": {
				"operationId": "UploadStudentPhoto",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Upload a profile photo for the student.\r\nSend as multipart/form-data with field name `photo`.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/users/student/me/cv": {
			"post": {
				"operationId": "UploadStudentCv",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Upload a CV document for the student.\r\nSend as multipart/form-data with field name `cv`.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/users/supervisor/me": {
			"get": {
				"operationId": "GetSupervisorProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get the authenticated supervisor's profile",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateSupervisorProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Update the authenticated supervisor's profile",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SupervisorProfileBody"
							}
						}
					}
				}
			}
		},
		"/api/users/company/me": {
			"get": {
				"operationId": "GetCompanyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get the authenticated company's profile",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateCompanyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Update the authenticated company's profile",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CompanyProfileBody"
							}
						}
					}
				}
			}
		},
		"/api/users/company/me/logo": {
			"post": {
				"operationId": "UploadCompanyLogo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Upload a company logo.\r\nSend as multipart/form-data with field name `logo`.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/users": {
			"get": {
				"operationId": "ListUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "List all users (Admin only)",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/users/{id}/toggle-active": {
			"patch": {
				"operationId": "ToggleUserActive",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Toggle a user's active status (Admin only)",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/messages": {
			"get": {
				"operationId": "GetConversations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get all conversations for the authenticated user",
				"tags": [
					"Messages"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/messages/start": {
			"post": {
				"operationId": "StartConversation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Start or retrieve an existing conversation with another user.\r\nReturns the conversation object (idempotent).",
				"tags": [
					"Messages"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StartConversationBody"
							}
						}
					}
				}
			}
		},
		"/api/messages/{conversationId}": {
			"get": {
				"operationId": "GetMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get all messages in a conversation",
				"tags": [
					"Messages"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "conversationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Send a message in a conversation.\r\nFile attachments are handled via multipart/form-data in the actual endpoint.\r\nReal-time delivery is handled via Socket.IO (`conversation:{id}` room).",
				"tags": [
					"Messages"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "conversationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendMessageBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}